FILS: Use AEAD cipher to protect EAPOL-Key frames (STA)
This adds AEAD counter values into the PTK data structure and modifies
wpa_eapol_key_send() to use AEAD cipher (AES-GCM for FILS AKMs) to
provide both integrity protection for the EAPOL-Key frame and encryption
for the Key Data field. It should be noted that this starts encrypting
the Key Data field in EAPOL-Key message 2/4 while it remains unencrypted
(but integrity protected) in non-FILS cases.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>